+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- ability to autofocus buttons ([#14](https://github.com/codewithkyle/notifyjs/issues/14))
+
+
 ## [2.0.3] - 2020-09-16
 
 ### Fixed